predictions with m(a) curves in notebook
