+++ bin/figures.py
@@ -7,10 +7,9 @@ import numpy as np
 from astropy import units as u
 from scipy.stats import spearmanr
 
-from relaxed import plot_defaults as plotdf
+from relaxed import plotting as rxplots
 from relaxed.correlations import get_am_corrs
 from relaxed.correlations import get_ma_corrs
-from relaxed.correlations import vol_jacknife_err
 from relaxed.cosmo import get_a_from_t
 from relaxed.cosmo import get_fractional_tdyn
 from relaxed.cosmo import get_t_from_a
@@ -29,93 +28,6 @@ plt.ioff()
 root = Path(__file__).absolute().parent.parent
 figsdir = root.joinpath("results", "figs")
 figsdir.mkdir(exist_ok=True, parents=False)
-
-
-def compute_metrics(pred_func, _x_test, _y_test, box_keep=None):
-    # pred_func should return ONLY 1 parameter. (e.g. lambda wrapped around indexing)
-    if box_keep is None:
-        box_keep = np.ones(_x_test.shape[0]).astype(bool)
-
-    x_test = _x_test[box_keep]
-    y_test = _y_test[box_keep].reshape(-1)
-
-    y_pred = pred_func(x_test).reshape(-1)
-    x = (y_pred - y_test) / np.std(y_test)  # normalize
-
-    return {
-        "mu": np.mean(x),
-        "sigma_ratio": np.std(y_pred) / np.std(y_test),
-        "spear": spearmanr(y_pred, y_test)[0],
-        "rscatter": np.std(x) / np.sqrt(2),
-        "mad": np.mean(np.abs(x)),
-    }
-
-
-def metrics_plot(
-    metrics_data: dict,
-    test_data: dict,
-    trained_models: dict,
-    cat_test,
-    params=("cvir",),
-    figsize=(12, 12),
-    nrows=2,
-    ncols=2,
-    ticksize=16,
-    bbox_to_anchor=(0.0, 1.0, 0.3, 0.3),
-    y_label_size=28,
-):
-    # NOTE: params and models MUST have same order
-    fig, axes = plt.subplots(nrows, ncols, figsize=figsize)
-    axes = axes.flatten()
-    model_metrics = {}
-
-    for jj, param in enumerate(params):
-        for mdl_name, (x_test, label, color, shape) in test_data.items():
-            model = trained_models[mdl_name]
-            pred_func = lambda x: model.predict(x)[:, jj]
-            y_test = cat_test[param].value.reshape(-1)
-            val_metrics = compute_metrics(pred_func, x_test, y_test)
-            errs = vol_jacknife_err(
-                cat_test, compute_metrics, pred_func, x_test, y_test, mode="dict"
-            )
-            model_metrics[(param, mdl_name)] = {k: (val_metrics[k], errs[k]) for k in metrics_data}
-
-    params_latex = [plotdf.latex_params[par] for par in params]
-    for ii, met in enumerate(metrics_data):
-        ax = axes[ii]
-        ax.set_xlim(-0.25, len(params))
-        if "yrange" in metrics_data[met]:
-            if metrics_data[met]["yrange"] is not None:
-                ax.set_ylim(metrics_data[met]["yrange"])
-        ax.set_xticks(np.array(list(range(len(params)))))
-        ax.set_xticklabels(params_latex)
-        ax.tick_params(axis="x", labelsize=ticksize)
-        x_bias = 0.0
-        for mdl_name, (x_test, label, color, marker) in test_data.items():
-            for jj, param in enumerate(params):
-                label = label if (jj == 0 and ii == 0) else None
-                val, err = model_metrics[(param, mdl_name)][met]
-                ax.errorbar(
-                    jj + x_bias,
-                    val,
-                    yerr=err,
-                    label=label,
-                    marker=marker,
-                    color=color,
-                    capsize=2.5,
-                    ms=6,
-                    capthick=2.0,
-                )
-            x_bias += 0.1
-
-        ax.set_ylabel(plotdf.latex_metrics[met], size=y_label_size)
-        if metrics_data[met].get("hline", None) is not None:
-            ax.axhline(metrics_data[met]["hline"], ls="--", color="k")
-
-        if ii == 0:
-            ax.legend(loc="lower left", bbox_to_anchor=bbox_to_anchor)
-    plt.tight_layout()
-    return fig
 
 
 def make_correlation_mah_plots():
@@ -143,8 +55,8 @@ def make_correlation_mah_plots():
     max_dict = {}
 
     for j, param in enumerate(params):
-        latex_param = plotdf.latex_params[param]
-        color = plotdf.cb_colors[j]
+        latex_param = rxplots.latex_params[param]
+        color = rxplots.cb_colors[j]
         corrs = get_ma_corrs(cat, param, ma)
         pos = corrs > 0
         neg = ~pos
@@ -165,7 +77,7 @@ def make_correlation_mah_plots():
     text = ""
     for j, param in enumerate(params):
         corr, scale = max_dict[param]
-        color = plotdf.cb_colors[j]
+        color = rxplots.cb_colors[j]
         ax.axvline(scale, linestyle="--", color=color)
         text += f"{param}: Max corr is {corr:.3f} at scale {scale:.3f}\n"
 
@@ -202,8 +114,8 @@ def make_correlation_mah_plots():
     max_dict = {}
 
     for j, param in enumerate(params):
-        latex_param = plotdf.latex_params[param]
-        color = plotdf.cb_colors[j]
+        latex_param = rxplots.latex_params[param]
+        color = rxplots.cb_colors[j]
         corrs = get_am_corrs(cat, param, am)
         pos = corrs >= 0
         neg = ~pos
@@ -222,7 +134,7 @@ def make_correlation_mah_plots():
     # draw a vertical line at max scales, output table.
     text = ""
     for j, param in enumerate(params):
-        color = plotdf.cb_colors[j]
+        color = rxplots.cb_colors[j]
         corr, mbin = max_dict[param]
         ax.axvline(mbin, linestyle="--", color=color)
         text += f"{param}: Max corr is {corr:.3f} at mass bin {mbin:.3f}\n"
@@ -254,7 +166,7 @@ def make_triangle(model_name, params, trained_models, datasets, sample_fn, figfi
     y2 = np.log(y_samples)
     y2[:, -3:] = np.exp(y2[:, -3:])
 
-    labels = [plotdf.latex_params[param] for param in params]
+    labels = [rxplots.latex_params[param] for param in params]
     fig = corner.corner(y1, labels=labels, max_n_ticks=3, color="C1", labelpad=0.05)
     fig = corner.corner(y2, labels=labels, max_n_ticks=3, fig=fig, color="C2", labelpad=0.05)
 
@@ -305,17 +217,17 @@ def make_triangle_plots():
 
     # (1) MultiGaussian sample triangle
     figfile = figsdir.joinpath("triangle_multigaussian.png")
-    sample_fn = lambda model, x: model.sample(x, 1).reshape(-1, model.n_targets)
+    sample_fn = lambda model, x: model.sample(x, 1).reshape(-1, model.n_targets)  # noqa: E731
     make_triangle("multicam", params, joint_models, datasets, sample_fn, figfile)
 
     # (2) LR pred triangle
     figfile = figsdir.joinpath("triangle_lr.png")
-    sample_fn = lambda model, x: model.predict(x).reshape(-1, model.n_targets)
+    sample_fn = lambda model, x: model.predict(x).reshape(-1, model.n_targets)  # noqa: E731
     make_triangle("multicam", params, joint_models, datasets, sample_fn, figfile)
 
     # (3) OptCAM pred triangle
     figfile = figsdir.joinpath("triangle_cam.png")
-    sample_fn = lambda model, x: model.predict(x).reshape(-1, model.n_targets)
+    sample_fn = lambda model, x: model.predict(x).reshape(-1, model.n_targets)  # noqa: E731
     make_triangle("optcam", params, joint_models, datasets, sample_fn, figfile)
 
 
@@ -552,25 +464,25 @@ def make_inv_pred_plots():
         "linear_all": (
             datasets["all"]["test"][0],
             r"\rm MultiCAM",
-            plotdf.cb_colors[0],
+            rxplots.cb_colors[0],
             markers[0],
         ),
         "linear_cvir": (
             datasets["cvir_only"]["test"][0],
             r"\rm MultiCAM $c_{\rm vir}$ only",
-            plotdf.cb_colors[1],
+            rxplots.cb_colors[1],
             markers[1],
         ),
         "linear_x0": (
             datasets["x0_only"]["test"][0],
             r"\rm MultiCAM $x_{\rm off}$ only",
-            plotdf.cb_colors[2],
+            rxplots.cb_colors[2],
             markers[2],
         ),
         "linear_tu": (
             datasets["tu_only"]["test"][0],
             r"\rm MultiCAM $T/\vert U \vert$ only",
-            plotdf.cb_colors[3],
+            rxplots.cb_colors[3],
             markers[3],
         ),
     }
@@ -742,23 +654,23 @@ def make_pred_plots():
 
     markers = ["o", "D", "s", "^"]
     test_data = {
-        "multicam": (datasets["all"]["test"][0], r"\rm MultiCAM", plotdf.cb_colors[0], markers[0]),
+        "multicam": (datasets["all"]["test"][0], r"\rm MultiCAM", rxplots.cb_colors[0], markers[0]),
         "multicam_alpha": (
             datasets["alpha"]["test"][0],
             r"\rm MultiCAM $\alpha$ only",
-            plotdf.cb_colors[1],
+            rxplots.cb_colors[1],
             markers[1],
         ),
         "multicam_diffmah_new": (
             datasets["diffmah_new"]["test"][0],
             r"\rm MultiCAM DiffMAH parameters with $a_{1/2}$",
-            plotdf.cb_colors[2],
+            rxplots.cb_colors[2],
             markers[2],
         ),
         "mixed_cam": (
             datasets["all"]["test"][0],
             r"\rm Optimal CAM",
-            plotdf.cb_colors[3],
+            rxplots.cb_colors[3],
             markers[3],
         ),
         # 'multicam_diffmah': (datasets['diffmah']['test'][0], r"\rm MultiCAM DiffMAH parameters", 'b', 's'),
